on patch_032: more solaris fixes
svn path=/branches/kpdf/annotations/kdegraphics/kpdf/; revision=424886

+++ ui/pageviewutils.cpp
@@ -445,13 +445,13 @@ void PageViewToolBar::buildToolBar()
     {
         myCols = 1 + (buttonsNumber * toolBarGridSize) /
                  (parentHeight - toolBarGridSize);
-        myRows = (int)ceilf( (float)buttonsNumber / (float)myCols );
+        myRows = (int)ceil( (float)buttonsNumber / (float)myCols );
     }
     else
     {
         myRows = 1 + (buttonsNumber * toolBarGridSize) /
                  (parentWidth - toolBarGridSize);
-        myCols = (int)ceilf( (float)buttonsNumber / (float)myRows );
+        myCols = (int)ceil( (float)buttonsNumber / (float)myRows );
     }
 
     // 2. compute widget size (from rows/cols)
